% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dinucmut.count}
\alias{dinucmut.count}
\title{Count of dinucleotide mutations (e.g. CC>AA, TT>CC)#'
This function counts the different types dinucleotide substitutions (e.g. CC to AA).}
\usage{
dinucmut.count(VCF)
}
\arguments{
\item{VCF}{a delimited file with columns "WT" and "MUT", containing values including "A", "T", "G" or "C" (a data.frame)}
}
\description{
Count of dinucleotide mutations (e.g. CC>AA, TT>CC)#'
This function counts the different types dinucleotide substitutions (e.g. CC to AA).
}
\examples{
res <- dinucmut.count(VCF)
}
\keyword{carlo,}
\keyword{monte}
\keyword{mutants,}
\keyword{mutations}
\keyword{point}
\keyword{substitutions,}

