% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dinucmut.mc}
\alias{dinucmut.mc}
\title{Dinucleotide Monte Carlo Simulation'
In Nik-Zainal et al. 2012 paper "Mutational Processes Molding the Genomes of 21 Breast Cancers", they note a higher rate of dinucleotide mutations than expected by chance. They describe this as estimated by Monte Carlo simulations. This function is my version of such a simulation. I scramble the existing data n=100 times, assuming the same number of mutations on each chromosoem Poisson distributed across the observed range e.g. between the first to last actual mutation. The function then calculates permuted counts (Expected Counts :E) for each dinucleotide type each time. The fraction of E > O (Observed Counts: O) is a Monte Carlo estimate of P (the null hypothesis).}
\usage{
dinucmut.mc(VCF, n.mc = 100)
}
\arguments{
\item{VCF}{a delimited file with columns "WT" and "MUT", containing values including "A", "T", "G" or "C" (a data.frame)}

\item{n.mc}{the number of Monte Carlo simulations (permutations) of the expected dincleotide rate E.}
}
\description{
Dinucleotide Monte Carlo Simulation'
In Nik-Zainal et al. 2012 paper "Mutational Processes Molding the Genomes of 21 Breast Cancers", they note a higher rate of dinucleotide mutations than expected by chance. They describe this as estimated by Monte Carlo simulations. This function is my version of such a simulation. I scramble the existing data n=100 times, assuming the same number of mutations on each chromosoem Poisson distributed across the observed range e.g. between the first to last actual mutation. The function then calculates permuted counts (Expected Counts :E) for each dinucleotide type each time. The fraction of E > O (Observed Counts: O) is a Monte Carlo estimate of P (the null hypothesis).
}
\examples{
res <- dinucmut.mc(VCF, n=100)
}
\keyword{carlo,}
\keyword{monte}
\keyword{mutants,}
\keyword{mutations}
\keyword{point}
\keyword{substitutions,}

