% Generated by roxygen2 (4.0.2): do not edit by hand
\name{expandseq}
\alias{expandseq}
\title{Get surrounding sequence context
uses the Bioconductor BSGenome framework to find}
\usage{
expandseq(VCF, start = "start.position", expand = 1)
}
\arguments{
\item{VCF}{a VCF tbl with columns containing at least "chr", "start.position", and "end.position" (a data.frame)}

\item{expand}{The numberof upstream and downstream bases to get (numeric, default = 1)}
}
\description{
Get surrounding sequence context
uses the Bioconductor BSGenome framework to find
}
\examples{
qqmutation(VCF, start= "start.position", around=1)
}
\keyword{Poisson,}
\keyword{kataegesis}
\keyword{mutants,}

