% Generated by roxygen2 (4.0.2): do not edit by hand
\name{imdensity.plot}
\alias{imdensity.plot}
\title{Intermutation Density Plot}
\usage{
imdensity.plot(VCF, chrN = NULL, facets = NULL, type = "density")
}
\arguments{
\item{VCF}{a VCF tbl with columns containing at least "chr", "start.position", and "end.position" (a data.frame)}
}
\description{
A density plot of intermutation distances with optional poissson fit
}
\examples{
imdensity.plot(VCF, chr=6)
}
\keyword{kataegesis}
\keyword{mutants,}
\keyword{rainfall,}

