% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nmf.plot}
\alias{nmf.plot}
\title{Non-negative matrix factorisation of mutation types
Alexandrov et al. pioneered the extraction of mutational signatures using the NMF technique}
\usage{
nmf.plot(count.df, rank = 4, method = "nsNMF", seed = "ica", nrun = 100,
  ...)
}
\arguments{
\item{count.df}{The format is a column named triplets (ATC, ATG, etc), a column named mutations(C, G, etc) and then columns of mutation counts for a set of samples}

\item{rank}{This is the number of mutation signatures you think maybe hiding within your data}

\item{method}{The NMF algorithm search algorithm. See the NMF package documentation for details of methods.}

\item{seed}{The NMF starting point guess. See the NMF package documentation for details of seeding.}
}
\description{
Non-negative matrix factorisation of mutation types
Alexandrov et al. pioneered the extraction of mutational signatures using the NMF technique
}
\examples{
nmf.plot(count.df, seed="ica")
}
\keyword{NMF,}
\keyword{factorisation,}
\keyword{kataegesis}
\keyword{matrix}
\keyword{mutations,}
\keyword{non-negative}

