% Generated by roxygen2 (4.0.2): do not edit by hand
\name{qqmutation}
\alias{qqmutation}
\title{QQ-Plots of mutation distribution vs theoretical random (Poisson) distribution}
\usage{
qqmutation(VCF, type = "unif")
}
\arguments{
\item{VCF}{a VCF tbl with columns containing at least "chr", "start.position", and "end.position" (a data.frame)}

\item{type}{One of "unif" or "qexp" (character, default = "unif")}
}
\description{
A density plot of intermutation distances with optional poissson fit
}
\examples{
qqmutation(VCF, type="qexp")
}
\keyword{Poisson,}
\keyword{kataegesis}
\keyword{mutants,}

