% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rainfall}
\alias{rainfall}
\title{Rainfall Plot}
\usage{
rainfall(VCF, chr, lower, upper, colour = NULL, facets = NULL,
  shape = NULL, bursts = F, s = 3, gamma = 3, vline = NULL, ...)
}
\arguments{
\item{VCF}{a VCF tbl with columns containing at least "chr", "start.position", and "end.position" (a data.frame)}

\item{chr}{the chromosome (character)}

\item{lower}{the lower range of the plot (numeric)}

\item{upper}{the upper range of the plot x-axis (numeric)}

\item{colour}{a column to specify point colours (character)}

\item{facets}{a column to specify plot faceting (character)}

\item{upper}{the upper range of the plot x-axis (numeric)}
}
\description{
A rainfall plot show the spatial clustering of mutations .. usually point mutations
}
\examples{
## before using the rainfall plot you might want to filter the VCF to contain mutations of a certain type.
## For instance this is filtering for APOBEC3A/B type point mutations
VCF <- filter(VCF, (WT=="C" & MUT=="T")|(WT=="C" & MUT=="G")|(WT=="G" & MUT=="A")|(WT=="G" & MUT=="C"))
## You might also add a column that can be used for shape or colour, e.g.
VCF <- mutate(VCF, mutation_type = ifelse((WT=="C" & MUT=="T")|(WT=="G" & MUT=="A"), "transition", "transversion"))
## Or guess  the strand of the mutation actual
VCF <- mutate(VCF, strand = ifelse((WT=="C" & MUT=="T")|(WT=="C" & MUT=="G"), "positive", "negative"))
rainfall(VCF, chr = 6, lower = 1.24e8, upper = 1.4e8, facets="strand", colour="mutation_type")
}
\keyword{kataegesis}
\keyword{mutants,}
\keyword{rainfall,}

