% Generated by roxygen2 (4.0.2): do not edit by hand
\name{renewal.plot}
\alias{renewal.plot}
\title{A renewal plot to detect non homogenous Poisson processes}
\usage{
renewal.plot(VCF, chr = 1)
}
\arguments{
\item{VCF}{}
}
\description{
This idea is borrowed from the analysis of spike train dat described in the STAR package - and reapplied here to a completely different problem. The basic idea is to use the
rank order of the intermutation distances OJ plotted against OJ+1 (the lag). This uses the plto space better than just plotting intermutaion vs intermutaion lag.
Homogenous Poisson processes fill the square  space randomly, deviations from homogeneity leave gaps.
}
\examples{
# a minimal example
VCF<-read.mutations(file="1317049.tsv", chr="Chromosome", start.position="Genome.start", end.position="Genome.stop")
renewal.plot(VCF)
}
\keyword{kataegesis}
\keyword{mutants,}
\keyword{rainfall,}

