% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dinucmut.count}
\alias{dinucmut.count}
\title{Frequency of dinucleotide mutations (e.g. CC>AA, TT>CC)}
\usage{
dinucmut.count(VCF)
}
\arguments{
\item{VCF}{a delimited file with columns containing at least "chr", "start.position", and "end.position" (a data.frame)}
}
\description{
In Nik-Zainal(2012) they describe a method to asees the numberof dinucleotide mutaions by first measuring the frequency of single point mutaion types (e.g A>C, T>C)
for each chromosome. These frequencies are then used to simulate many 1000s of  times the expected dinucleotide frequencies given
}
\examples{
VCF <- read.mutations(file="1317049.tsv", chr="Chromosome", start.position="Genome.start", end.position="Genome.stop", strand= "Strand", WT="WT.seq", MUT="Mut.seq", sampleID="Sample.Name")
res <- dinuc.freq.table(VCF)
}
\keyword{carlo,}
\keyword{dinucleotide}
\keyword{monte}
\keyword{mutants,}

